+++ CS320_Team_Fractal_Website/src/edu/ycp/cs320/CS320_Team_Fractal_Website/model/account/Admin.java
@@ -16,6 +16,8 @@ public class Admin extends User{
 	
 	@Override
 	public boolean deleteFractal(int fractalId, IDatabase db){
+		if(db == null) return false;
+		
 		db.deleteFractal(fractalId);
 		
 		return true;
